feat: persist-item 도메인 생성자에 id 추가

+++ persistence/persist-item/src/main/java/store/nightmarket/persistence/persistitem/entity/valueobject/Rating.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode
 public class Rating {
 
-    private int value;
+    private float value;
 
 }
